Add mutest module
(think about dropping it if ever it becomes part of the SDK)

So actually, we don't even run the unit tests. We could by adding
`run-tests: true`, however the unit tests then fail, in particular the
appstream test fails, as it requires network access.

So, let's keep test disabled for now, and let's enable it in the next
version.

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>