Add --socket=fallback-x11 and --device=dri
It's unclear if it's needed or not. Does a GTK app make use of OpenGL?
I'd say yes, but that doesn't make me 100% certain that --device=dri is
needed.

In any case, reading the doc again at [1], these are standard
permissions, meaning it's safe to list them all, won't hurt.

[1]: <https://docs.flatpak.org/en/latest/sandbox-permissions.html#permissions-guidelines>

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>